Working StartService for utitan-input

+++ app/src/main/java/com/rumplestilzken/gargoylesettings/touchpad/TouchpadProcessor.java
@@ -9,12 +9,16 @@ import com.rumplestilzken.gargoylesettings.provider.RootProvider;
 import com.rumplestilzken.gargoylesettings.service.ServiceProcessor;
 
 import java.io.DataOutputStream;
+import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
+import java.nio.charset.StandardCharsets;
+
+import android.util.Base64;
 
 public class TouchpadProcessor {
     private static Context context = null;
@@ -26,42 +30,33 @@ public class TouchpadProcessor {
         return context;
     }
 
-    public static void enableTouchpadScrolling() {
+    public static void enableTouchpadScrolling(boolean service) {
         Log.d("TouchpadProcessor", "Starting uinput_titan");
         ServiceProcessor.startOrStop(true, "uninput_titan");
-        RebootProvider.setRequiresReboot(true);
-        RootProvider.Remount();
-        RootProvider.RunAsRoot("rm /etc/excluded-input-devices.xml");
+        if(!service)
+        {
+            RebootProvider.setRequiresReboot(true);
+            RootProvider.Remount();
+            RootProvider.RunAsRoot("rm /etc/excluded-input-devices.xml");
+        }
     }
 
-    public static void disableTouchpadScrolling() {
+    public static void disableTouchpadScrolling(boolean service) {
        Log.d("TouchpadProcessor", "Stopping uinput_titan");
        ServiceProcessor.startOrStop(false, "uninput_titan");
-       RebootProvider.setRequiresReboot(true);
-       RootProvider.Remount();
-       InputStream in = context.getResources().openRawResource(R.raw.excluded_input_devices);
-       FileOutputStream out = null;
-       try {
-           out = new FileOutputStream("/etc/excluded-input-devices.xml");
-       } catch (FileNotFoundException e) {
-           throw new RuntimeException(e);
-       }
-       byte[] buff = new byte[1024];
-       int read = 0;
-       try {
-           while ((read = in.read(buff)) > 0) {
-               out.write(buff, 0, read);
-           }
-       } catch (IOException e) {
-           throw new RuntimeException(e);
-       } finally {
+       if(!service) {
+           RebootProvider.setRequiresReboot(true);
+           RootProvider.Remount();
+           RootProvider.RunAsRoot("touch /etc/excluded-input-devices.xml");
            try {
-               in.close();
-               out.close();
+               if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.TIRAMISU) {
+                   byte[] bytes = context.getResources().openRawResource(R.raw.excluded_input_devices).readAllBytes();
+                   String string = new String(bytes, StandardCharsets.UTF_8);
+                   RootProvider.RunAsRoot("echo '" + string + "' > /etc/excluded-input-devices.xml");
+               }
            } catch (IOException e) {
                throw new RuntimeException(e);
            }
-
        }
     }
